refactor(pnpr): extract pnpr-config
Extracting `config` was blocked by a cycle: `config` reached into `s3` for
`S3Settings` and `build_s3_store`, `s3` into `storage`, and `storage` back
into `config`.

Neither of those items belonged in `s3`. `S3Settings` is the deserialized
YAML `s3:` block — config's own data, filed under the backend that consumes
it. And `build_s3_store` builds an `AmazonS3Builder` from that block; it
touches `object_store` but nothing in our `s3` module, so it sits just as well
next to the settings it reads. Moving both leaves `config` with no dependency
on `s3` at all and the cycle gone, with no behavior change: the store is still
built once at config-load time, which is what makes constructing the storage
layer infallible.

`config` is then a leaf, so it becomes `pnpr-config`, carrying `config.yaml`
(the bundled default it `include_str!`s) with it.

Two items the boundary caught as reaching across: `RedactedHeaders` and
`UpstreamConfig::with_defaults`, both now `pub`. `with_defaults` also gained
`#[must_use]`, the same public-API lint consequence the last extraction saw.

The suite is unchanged at 714, of which 144 now run under `pnpr-config`.
