+++ pnpr/crates/pnpr/Cargo.toml
@@ -25,6 +25,7 @@ backend-mysql    = ["dep:sqlx", "sqlx/mysql", "pnpr-error/backend-mysql"]
 backend-postgres = ["dep:sqlx", "sqlx/postgres", "pnpr-error/backend-postgres"]
 
 [dependencies]
+pnpr-config                   = { workspace = true }
 pnpr-error                    = { workspace = true }
 pnpr-package-name             = { workspace = true }
 pnpr-policy                   = { workspace = true }
