+++ pnpr/crates/pnpr/src/server/routing.rs
@@ -20,7 +20,8 @@ use tower_http::{
 };
 use tracing::Span;
 
-use crate::{auth::AuthState, config::Config, storage::Storage, upstream::Upstream};
+use crate::{auth::AuthState, storage::Storage, upstream::Upstream};
+use pnpr_config::Config;
 
 use super::{
     AppInner, AppState, AuthedCaller, MAX_ARTIFACT_BLOB_BODY_BYTES,
