+++ pnpr/crates/pnpr/src/storage.rs
@@ -1,7 +1,8 @@
-use crate::{config::HostedStoreConfig, s3::S3Store, streaming};
+use crate::{s3::S3Store, streaming};
 use async_trait::async_trait;
 use axum::body::Body;
 use pnpm_crypto_hash::integrity_addressed_tarball_integrity;
+use pnpr_config::HostedStoreConfig;
 use pnpr_error::{RegistryError, Result};
 use pnpr_package_name::PackageName;
 use serde::{Deserialize, Serialize};
