+++ pnpr/crates/pnpr/src/streaming/tests.rs
@@ -1,6 +1,7 @@
 use super::{TarballStreamError, integrity_checker, parse_integrity, stream_verified_to_cache};
-use crate::{config::HostedStoreConfig, storage::Storage};
+use crate::storage::Storage;
 use futures_util::StreamExt;
+use pnpr_config::HostedStoreConfig;
 use pnpr_package_name::PackageName;
 use ssri::{Algorithm, Integrity, IntegrityOpts};
 use std::{path::Path, sync::Arc, time::Duration};
